moved shared frontend helpers into modules
