+++ frontend/src/appInfo.js
@@ -1,0 +1,3 @@
+import packageInfo from "../../package.json";
+
+export const APP_VERSION = packageInfo.version;
